+++ app/controllers/dashboard/user_sessions_controller.rb
@@ -2,7 +2,7 @@ module Dashboard
   class UserSessionsController < ApplicationController
     def destroy
       logout
-      redirect_to login_path, notice: t('.logout')
+      redirect_to home_login_path, notice: t('.logout')
     end
   end
 end
