--- app/controllers/user_sessions_controller.rb
@@ -1,1 +1,0 @@
-
