+++ app/views/dashboard/cards/_card.html.erb
@@ -3,8 +3,8 @@
   <td><%= card.original_text %></td>
   <td><%= card.translated_text %></td>
   <td><%= card.review_date.strftime("%d/%m/%Y") %></td>
-  <td><%= link_to 'Edit', edit_card_path(card) %></td>
-  <td><%= link_to 'Delete', card_path(card),
+  <td><%= link_to 'Edit', edit_dashboard_card_path(card) %></td>
+  <td><%= link_to 'Delete', dashboard_card_path(card),
                   method: :delete,
                   remote: true,
                   data: { confirm: 'Are you sure?' } %></td>
