+++ app/views/dashboard/cards/_form.html.erb
@@ -1,4 +1,4 @@
-<%= simple_form_for @card do |f| %>
+<%= simple_form_for [:dashboard, @card] do |f| %>
     <%= f.input :deck_id, collection: @decks %>
     <%= f.input :original_text %>
     <%= f.input :translated_text %>
