+++ app/views/dashboard/cards/new.html.erb
