+++ app/views/home/user_sessions/_form.html.erb
@@ -1,4 +1,4 @@
-<%= simple_form_for :user_session, :url => user_sessions_path do |f| %>
+<%= simple_form_for :user_session, :url => home_user_sessions_path do |f| %>
   <%= f.input :email %>
   <%= f.input :password %>
   <%= f.button :submit, t('.enter') %>
