+++ app/views/layouts/_header.html.erb
@@ -9,10 +9,10 @@
           <li class="dropdown">
             <%= link_to t('.my_cards'), '#', class: 'dropdown-toggle', 'data-toggle' => 'dropdown' %>
             <ul class="dropdown-menu">
-              <li><%= link_to t('.decks'), decks_path %></li>
-              <li><%= link_to t('.add_deck'), new_deck_path %></li>
-              <li><%= link_to t('.cards'), cards_path %></li>
-              <li><%= link_to t('.add_card'), new_card_path %></li>
+              <li><%= link_to t('.decks'), dashboard_decks_path %></li>
+              <li><%= link_to t('.add_deck'), new_dashboard_deck_path %></li>
+              <li><%= link_to t('.cards'), dashboard_cards_path %></li>
+              <li><%= link_to t('.add_card'), new_dashboard_card_path %></li>
             </ul>
           </li>
           <li class="dropdown">
